requirements_dev.txt
pip==22.2.2
bump2version==1.0.1
wheel==0.37.1
watchdog==2.1.9
flake8==5.0.4
tox==3.26.0
coverage==6.5.0
Sphinx==5.2.3
twine==4.0.1
Click==8.1.3
pytest==7.1.3
black==22.10.0
pyhumps==3.7.3
pydantic==1.10.2
